+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - PNDA-3345: Provide the app_packages HDFS location (from Pillar) to applications deployed with DM
 - PNDA-3548: Upgrade Kafka manager to version 1.3.3.15
 - PNDA-3527: Add dev/prod queues to YARN CDH config.
+- PNDA-3528: Add some pillars for the resource manager and used in the dm-config.
 
 ### Changed
 - PNDA-3545: Configure Hive and Hive2 Ambari views to run as the hdfs super user